@@ -1,15 +1,15 @@
+
+import {first} from 'rxjs/operators';
 import { InjectionToken, NgZone } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
+import { Subscription ,  Observable ,  Subscriber } from 'rxjs';
+import { observeOn } from 'rxjs/operators';
 import { queue } from 'rxjs/scheduler/queue';
 
 import firebase from '@firebase/app';
 import { FirebaseApp, FirebaseOptions } from '@firebase/app-types';
 
 import 'zone.js';
-import 'rxjs/add/operator/first';
-import { Subscriber } from 'rxjs/Subscriber';
-import { observeOn } from 'rxjs/operator/observeOn';
+
 
 export const FirebaseAppName = new InjectionToken<string>('angularfire2.appName');
 export const FirebaseAppConfig = new InjectionToken<FirebaseOptions>('angularfire2.config');
@@ -27,7 +27,7 @@ export class FirebaseZoneScheduler {
     return new Observable<T>(subscriber => {
       const noop = () => {};
       const task = Zone.current.scheduleMacroTask('firebaseZoneBlock', noop, {}, noop, noop);
-      obs$.first().subscribe(() => this.zone.runOutsideAngular(() => task.invoke()));
+      obs$.pipe(first()).subscribe(() => this.zone.runOutsideAngular(() => task.invoke()));
       return obs$.subscribe(subscriber);
     });
   }
@@ -37,6 +37,6 @@ export class FirebaseZoneScheduler {
         return obs$.subscribe(subscriber);
       });
     });
-    return observeOn.call(outsideAngular, this);
+    return outsideAngular;
   }
 }
