@@ -1,7 +1,7 @@
 import { FirebaseApp, FirebaseAppConfig, AngularFireModule} from 'angularfire2';
 import { AngularFireDatabase, AngularFireDatabaseModule, FirebaseListObservable, FirebaseObjectFactory } from 'angularfire2/database-deprecated';
-import { Observer } from 'rxjs/Observer';
-import { map } from 'rxjs/operator/map';
+import { Observer } from 'rxjs';
+
 import { FirebaseApp as FBApp } from '@firebase/app-types';
 import { DataSnapshot, Reference } from '@firebase/database-types';
 import { unwrapMapFn } from './utils';
