@@ -1,9 +1,10 @@
+
+import {filter, map, scan} from 'rxjs/operators';
 import { DocumentChangeType, CollectionReference, Query, DocumentReference } from '@firebase/firestore-types';
-import { Observable } from 'rxjs/Observable';
-import { Subscriber } from 'rxjs/Subscriber';
+import { Observable ,  Subscriber } from 'rxjs';
 import { fromCollectionRef } from '../observable/fromRef';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/filter';
+
+
 
 import { Injectable } from '@angular/core';
 
@@ -12,9 +13,9 @@ import { docChanges, sortedChanges } from './changes';
 import { AngularFirestoreDocument } from '../document/document';
 import { AngularFirestore } from '../firestore';
 
-import { observeOn } from 'rxjs/operator/observeOn';
 
-import 'rxjs/add/observable/of';
+
+
 
 export function validateEventsArray(events?: DocumentChangeType[]) {
   if(!events || events!.length === 0) {
@@ -72,9 +73,9 @@ export class AngularFirestoreCollection<T> {
     if(!events || events.length === 0) {
       return docChanges(this.query);
     }
-    return this.afs.scheduler.keepUnstableUntilFirst(docChanges(this.query)
-      .map(actions => actions.filter(change => events.indexOf(change.type) > -1))
-      .filter(changes =>  changes.length > 0));
+    return this.afs.scheduler.keepUnstableUntilFirst(docChanges(this.query).pipe(
+      map(actions => actions.filter(change => events.indexOf(change.type) > -1)),
+      filter(changes =>  changes.length > 0),));
   }
 
   /**
@@ -83,7 +84,7 @@ export class AngularFirestoreCollection<T> {
    * @param events
    */
   auditTrail(events?: DocumentChangeType[]): Observable<DocumentChangeAction[]> {
-    return this.stateChanges(events).scan((current, action) => [...current, ...action], []);
+    return this.stateChanges(events).pipe(scan((current, action) => [...current, ...action], []));
   }
 
   /**
@@ -104,8 +105,8 @@ export class AngularFirestoreCollection<T> {
   valueChanges(): Observable<T[]> {
     const fromCollectionRef$ = fromCollectionRef(this.query);
     const scheduled$ = this.afs.scheduler.runOutsideAngular(fromCollectionRef$);
-    return this.afs.scheduler.keepUnstableUntilFirst(scheduled$)
-      .map(actions => actions.payload.docs.map(a => a.data()) as T[]);
+    return this.afs.scheduler.keepUnstableUntilFirst(scheduled$).pipe(
+      map(actions => actions.payload.docs.map(a => a.data()) as T[]));
   }
 
   /**
