@@ -1,23 +1,26 @@
+
+import {merge as observableMerge, of as observableOf,  Observable } from 'rxjs';
+
+import {scan, switchMap, distinctUntilChanged} from 'rxjs/operators';
 import { fromRef } from '../observable/fromRef';
-import { Observable } from 'rxjs/Observable';
 import { DatabaseQuery, ChildEvent, AngularFireAction, SnapshotAction } from '../interfaces';
 import { isNil } from '../utils';
 
-import 'rxjs/add/operator/scan';
-import 'rxjs/add/observable/merge';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/filter';
-import 'rxjs/add/operator/delay';
-import 'rxjs/add/operator/distinctUntilChanged';
+
+
+
+
+
+
+
 
 export function listChanges<T>(ref: DatabaseQuery, events: ChildEvent[]): Observable<SnapshotAction[]> {
-  return fromRef(ref, 'value', 'once').switchMap(snapshotAction => {
-    const childEvent$ = [Observable.of(snapshotAction)];
+  return fromRef(ref, 'value', 'once').pipe(switchMap(snapshotAction => {
+    const childEvent$ = [observableOf(snapshotAction)];
     events.forEach(event => childEvent$.push(fromRef(ref, event)));
-    return Observable.merge(...childEvent$).scan(buildView, [])
-  })
-  .distinctUntilChanged();
+    return observableMerge(...childEvent$).pipe(scan(buildView, []))
+  }),
+  distinctUntilChanged(),);
 }
 
 function positionFor(changes: SnapshotAction[], key) {
