@@ -1,9 +1,10 @@
+
+import {merge as observableMerge,  Observable } from 'rxjs';
 import { DatabaseQuery, ChildEvent, AngularFireAction, SnapshotAction } from '../interfaces';
 import { fromRef } from '../observable/fromRef';
 import { validateEventsArray } from './utils';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/merge';
-import 'rxjs/add/operator/scan';
+
+
 import { DataSnapshot } from '@firebase/database-types';
 import { AngularFireDatabase } from '../database';
 
@@ -14,5 +15,5 @@ export function createStateChanges(query: DatabaseQuery, afDatabase: AngularFire
 export function stateChanges(query: DatabaseQuery, events?: ChildEvent[]) {
   events = validateEventsArray(events)!;
   const childEvent$ = events.map(event => fromRef(query, event));
-  return Observable.merge(...childEvent$);
+  return observableMerge(...childEvent$);
 }
